Fix timeout issue with websockets.
Nginx has a setting called proxy_read_timeout that defaults to 60s.
This is a problem for our web sockets, because we may want to keep
these open for weeks, months, or even years. The solution is to
implement a timer that sends a websocket ping command every once in
a while. I've chosen 15s here, which should be low enough to reset
any intermediate proxy's timeout.