Rework mongo tailing mechanism
The previous mongo tailing mechanism wasn't very effective when using
a custom query. I've reworked it a bit to play nicely with custom
queries, as is the case with Avatar (which has a fairly complex
conditional query to only obtain crawl data relevant to one of the
URLs that it knows about).

Also: add a hacky way to encode ObjectId's in the tail query. I
don't think that a mongo query containing an ObjectId can be serialized
to JSON, and the work around requires importing the entire PyMongo
module -- a hideous implementation detail to expose to the client.
Therefore, the hack I chose is to recursively visit all nested values
in the query and check if the match the string ObjectId(...), and if so,
convert them to an actual ObjectId just before querying.